add MessageDetails screen components: include AudioPlayer, AudioWave, MessageBubble, MessageContent, MessageMeta, and ProfileImage
